requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.14.3
httpx==0.27.0
tiktoken==0.7.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
rank-bm25==0.2.2
sse-starlette==1.8.2